Add support for more gates types
This commit adds support for NAND, NOR, and XNOR gates.
It can completely read in a trust-hub benchmark file. However,
it prunes keys by iterating over the entire keyspace. This
needs to be converted to adding a clause to the SAT solver that
says "if this is the input, this must be the output" and then
it will not suggest the wrong keys again.